♻️  refactor mission files
