+++ suave_missions/suave_missions/time_constrained_mission.py
@@ -7,18 +7,23 @@ from datetime import datetime
 from rclpy.callback_groups import MutuallyExclusiveCallbackGroup
 from rclpy.executors import MultiThreadedExecutor
 from std_msgs.msg import Float32
-from suave_missions.mission_planner import MissionPlanner
+from suave_missions.inspection_mission import InspectionMission
 
 
-class MissionTimeConstrained(MissionPlanner):
+class MissionTimeConstrained(InspectionMission):
     def __init__(self, node_name='time_contrained_mission_node'):
         super().__init__(node_name)
-        self.get_logger().info('New log')
+        if self.result_filename == 'mission_results':
+            self.result_filename = 'time_mission_results'
 
         self.mission_name = 'time constrained ' + str(self.adaptation_manager)
         self.metrics_header = [
-            'mission_name', 'datetime', 'initial pos (x,y)', 'time budget (s)',
-            'time search (s)', 'distance inspected (m)']
+            'mission_name',
+            'datetime',
+            'initial pos (x,y)',
+            'time budget (s)',
+            'time search (s)',
+            'distance inspected (m)']
 
         self.declare_parameter('time_limit', 300)
         self.time_limit = self.get_parameter('time_limit').value
@@ -30,64 +35,45 @@ class MissionTimeConstrained(MissionPlanner):
             1,
             callback_group=MutuallyExclusiveCallbackGroup())
 
-        self.mission_start_time = None
-        self.pipeline_detected_time = None
         self.distance_inspected = 0.0
-
-    def distance_inspected_cb(self, msg):
-        self.distance_inspected = msg.data
-
-    def time_monitor_cb(self):
-        current_time = self.get_clock().now()
-        elapsed_time = current_time - self.mission_start_time
-        if elapsed_time.to_msg().sec >= self.time_limit:
-            self.abort_mission = True
-            detection_time_delta = -1
-            if self.pipeline_detected_time is not None:
-                detection_time_delta = \
-                    self.pipeline_detected_time - self.mission_start_time
-                detection_time_delta = detection_time_delta.to_msg().sec
-
-            # TODO: get this automatically
-            self.initial_x = 1.0
-            self.initial_y = 1.0
-            mission_metrics = [
-                self.mission_name,
-                datetime.now().strftime("%b-%d-%Y-%H-%M-%S"),
-                '({0}, {1})'.format(
-                    round(self.initial_x, 2), round(self.initial_y, 2)),
-                self.time_limit,
-                detection_time_delta,
-                self.distance_inspected]
-            self.save_metrics(mission_metrics)
-            self.get_logger().info('Time limited reached. Mission aborted!')
-            os.system("touch ~/suave_ws/mission.done")
-            self.time_monitor_timer.destroy()
-
-    def perform_mission(self):
-        self.get_logger().info("Pipeline inspection mission starting!!")
-        self.timer = self.create_rate(1)
-
-        while not self.status.armed:
-            self.get_logger().info(
-                'BlueROV is armed: {}'.format(self.status.armed))
-            self.arm_motors(True)
-            self.timer.sleep()
-
-        guided_mode = 'GUIDED'
-        while self.status.mode != guided_mode:
-            self.get_logger().info(
-                'BlueROV mode is : {} waiting for GUIDED mode'.format(
-                 self.status.mode))
-            self.set_mode(guided_mode)
-            self.timer.sleep()
-
-        self.mission_start_time = self.get_clock().now()
         self.time_monitor_timer = self.create_timer(0.5, self.time_monitor_cb)
 
-        self.perform_task('search_pipeline', lambda: self.pipeline_detected)
-        self.perform_task('inspect_pipeline', lambda: self.pipeline_inspected)
+    def distance_inspected_cb(self, msg):
+        if self.abort_mission is False:
+            self.distance_inspected = msg.data
 
+    def time_monitor_cb(self):
+        if self.mission_start_time is not None:
+            current_time = self.get_clock().now()
+            elapsed_time = current_time - self.mission_start_time
+            if elapsed_time.to_msg().sec >= self.time_limit:
+                self.abort_mission = True
+                self.time_monitor_timer.destroy()
+
+    def exit_mission(self):
+        detection_time_delta = -1
+        if self.pipeline_detected_time is not None:
+            detection_time_delta = \
+                self.pipeline_detected_time - self.mission_start_time
+            detection_time_delta = detection_time_delta.to_msg().sec
+
+        self.get_logger().info(
+            'Time elapsed to detect pipeline: {} seconds'.format(
+                detection_time_delta))
+        self.get_logger().info(
+            'Distance inspected: {} meters'.format(
+                self.distance_inspected))
+
+        mission_metrics = [
+            self.mission_name,
+            datetime.now().strftime("%b-%d-%Y-%H-%M-%S"),
+            '({0}, {1})'.format(
+                round(self.initial_x, 2), round(self.initial_y, 2)),
+            self.time_limit,
+            detection_time_delta,
+            self.distance_inspected]
+
+        self.save_metrics(mission_metrics)
         os.system("touch ~/suave_ws/mission.done")
